backend/requirements.txt
flask==3.0.0
flask-cors==4.0.1
opencv-python==4.8.1.78
deepface==0.0.96
numpy==1.26.4
pillow==10.2.0
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0

